% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model.r
\name{acousticModel}
\alias{acousticModel}
\title{Base class for Acoustic Telemetry Models}
\description{
An R6 base class that is used to fit a Markov modulated Poisson process to acoustic telemetry data.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{detectors}}{Dataframe of detector locations and states.}

\item{\code{studyperiod}}{Study period stored for convenience.}

\item{\code{statespace}}{Dataframe of state space centroids, indices, and their neighbourhood structure.}

\item{\code{nstates}}{Number of physical location state (excluding mortality state).}

\item{\code{resolution}}{Width of grid cell in a state.}

\item{\code{xymin}}{Lower corner of the statespace.}

\item{\code{itoj}}{State i to j indices for building the generator matrix.}

\item{\code{design_ou}}{Matrix that shows direction coordinates and x,y values at same scale as itoj.}

\item{\code{designmatrix}}{Design matrix for advection on itoj for building generator matrix.}

\item{\code{absorbingstates}}{Holds which states are physically holding if they are supplied.}

\item{\code{sim_obs}}{Simulated observations when running a simulation.}

\item{\code{sim_move}}{Simulated true movement when running a simulation.}

\item{\code{observations}}{Processed observations ready for analysis.}

\item{\code{emissionrate}}{Rate that the tags phyiscally ping.}

\item{\code{pars}}{initial values for fitting the model.}

\item{\code{negll}}{Negative log likelihood RTMB object.}

\item{\code{gr_negll}}{Gradient of the negative log likelihood.}

\item{\code{lookup_animal_id}}{Lookup matrix for matching animal id.}

\item{\code{lookup_detector_id}}{Lookup matrix for matching detectors.}

\item{\code{lookup_statespace}}{Lookup matrix for matching state space and locations.}

\item{\code{estimated_pars}}{List of estimated parameters.}

\item{\code{delta_xy}}{Distance between states pairwise.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-acousticModel-new}{\code{acousticModel$new()}}
\item \href{#method-acousticModel-processStatesDetectors}{\code{acousticModel$processStatesDetectors()}}
\item \href{#method-acousticModel-modelSetUp}{\code{acousticModel$modelSetUp()}}
\item \href{#method-acousticModel-simulate}{\code{acousticModel$simulate()}}
\item \href{#method-acousticModel-makeADFun}{\code{acousticModel$makeADFun()}}
\item \href{#method-acousticModel-solveExpectedPath}{\code{acousticModel$solveExpectedPath()}}
\item \href{#method-acousticModel-calculateQ}{\code{acousticModel$calculateQ()}}
\item \href{#method-acousticModel-processData}{\code{acousticModel$processData()}}
\item \href{#method-acousticModel-simFit}{\code{acousticModel$simFit()}}
\item \href{#method-acousticModel-fitMove}{\code{acousticModel$fitMove()}}
\item \href{#method-acousticModel-buildModel}{\code{acousticModel$buildModel()}}
\item \href{#method-acousticModel-clone}{\code{acousticModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-acousticModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-acousticModel-new}{}}}
\subsection{Method \code{new()}}{
Initialize the R6 object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{acousticModel$new(grid, detectors, formula = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{grid}}{State space that represents where the animal can move and any covariates of interest, must contain x,y columns.}

\item{\code{detectors}}{Dataframe of detector locations, must contain columns x,y.}

\item{\code{formula}}{Formula for covariates to relate statespace to advection (optional).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-acousticModel-processStatesDetectors"></a>}}
\if{latex}{\out{\hypertarget{method-acousticModel-processStatesDetectors}{}}}
\subsection{Method \code{processStatesDetectors()}}{
Process Statespace and Detectors
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{acousticModel$processStatesDetectors(statespace, detectors)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{statespace}}{Dataframe that represents where the animal can move and any covariates of interest, must contain x,y columns.}

\item{\code{detectors}}{Dataframe of detector locations, must contain columns x,y.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create indices to the state space and match with detectors. Also computes match Left, Right, Up, Down moves. Output is stored in R6 object.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-acousticModel-modelSetUp"></a>}}
\if{latex}{\out{\hypertarget{method-acousticModel-modelSetUp}{}}}
\subsection{Method \code{modelSetUp()}}{
Set up advection covariates
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{acousticModel$modelSetUp(formula = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{formula}}{Formula to relate glm like relationship of advection to the covariates.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Uses central differences to compute the local gradient of each covariate to include in advection model.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-acousticModel-simulate"></a>}}
\if{latex}{\out{\hypertarget{method-acousticModel-simulate}{}}}
\subsection{Method \code{simulate()}}{
Simulate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{acousticModel$simulate(
  N = 1,
  alpha = 0,
  beta = NULL,
  q = NULL,
  gamma = NULL,
  formula = NULL,
  emissionrate = 30,
  studyperiod = NULL,
  absorbingstates = NULL,
  startbbox = NULL,
  mu = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{N}}{Number of animals.}

\item{\code{alpha}}{Diffusion rate.}

\item{\code{beta}}{Advection rates.}

\item{\code{q}}{Detection probability.}

\item{\code{gamma}}{Centre of attraction.}

\item{\code{formula}}{Formula to relate covariates to advection (optional).}

\item{\code{emissionrate}}{Rate that the tags send a signal for detection.}

\item{\code{studyperiod}}{Time period of the study.}

\item{\code{absorbingstates}}{Vector of states indicating which locations the animal will not leave if enters.}

\item{\code{startbbox}}{Vector indicating an area that the animal taggging is taking place.}

\item{\code{mu}}{Rate of mortality per unit time.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Simulate from the Markov modulated Poisson process data that is equivalent to the acoustic telemetry study using diffusion-advection for movement. See \code{?simulate_mmpp} for more details.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-acousticModel-makeADFun"></a>}}
\if{latex}{\out{\hypertarget{method-acousticModel-makeADFun}{}}}
\subsection{Method \code{makeADFun()}}{
Make negative log likelihood RTMB object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{acousticModel$makeADFun(
  alpha,
  beta,
  q,
  mu = NULL,
  gamma = NULL,
  emissionrate = 720,
  studyperiod = NULL,
  control = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{Initial diffusion rate.}

\item{\code{beta}}{Initial advection rates.}

\item{\code{q}}{Initial detection probability.}

\item{\code{mu}}{Initial mortality rate.}

\item{\code{gamma}}{Initial centre of attraction.}

\item{\code{emissionrate}}{Rate that the tags send a signal for detection.}

\item{\code{studyperiod}}{Time period of the study.}

\item{\code{control}}{List of object controls which include tolerance, rescale_freq, Nmax, uniformization, trace that controls the RTMB function \code{expAv}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Simulate from the Markov modulated Poisson process data that is equivalent to the acoustic telemetry study using diffusion-advection for movement. See \code{?simulate_mmpp} for more details.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-acousticModel-solveExpectedPath"></a>}}
\if{latex}{\out{\hypertarget{method-acousticModel-solveExpectedPath}{}}}
\subsection{Method \code{solveExpectedPath()}}{
Solve Expected Paths for Observed Detections
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{acousticModel$solveExpectedPath(
  alpha,
  beta,
  q,
  mu = NULL,
  gamma = NULL,
  tstart = 0,
  tend = NULL,
  delta_t = NULL,
  control = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{Initial diffusion rate.}

\item{\code{beta}}{Initial advection rates.}

\item{\code{q}}{Initial detection probability.}

\item{\code{mu}}{Initial mortality rate.}

\item{\code{gamma}}{Initial centre of attraction.}

\item{\code{tstart}}{Start time of solving.}

\item{\code{tend}}{Time end of solving.}

\item{\code{delta_t}}{Change in time to step through.}

\item{\code{control}}{List of object controls which include tolerance, rescale_freq, Nmax, uniformization, trace that controls the RTMB function \code{expAv}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Run forward and backward algorithm to compute probability of where the animals were seen. Then use the Vertibi algorithm to estimate the most likely path the animal takes.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-acousticModel-calculateQ"></a>}}
\if{latex}{\out{\hypertarget{method-acousticModel-calculateQ}{}}}
\subsection{Method \code{calculateQ()}}{
Compute the Movement Generator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{acousticModel$calculateQ(alpha, beta, mu = NULL, gamma = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{Diffusion rate.}

\item{\code{beta}}{Advection rates.}

\item{\code{mu}}{Mortality rate.}

\item{\code{gamma}}{Centre of attraction.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-acousticModel-processData"></a>}}
\if{latex}{\out{\hypertarget{method-acousticModel-processData}{}}}
\subsection{Method \code{processData()}}{
Process either data from real observations or from the simulation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{acousticModel$processData(
  detections = NULL,
  releases = NULL,
  known_fates = NULL,
  absorbing_states = NULL,
  studyperiod = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{detections}}{Dataframe for when and where each animal detection occurs.}

\item{\code{releases}}{Where and when the animals were released.}

\item{\code{known_fates}}{Time and states of known fates of animals.}

\item{\code{absorbing_states}}{States that are absorbing if the animal arrives.}

\item{\code{studyperiod}}{How long was the study run for.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
See \code{?process_data} for details.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-acousticModel-simFit"></a>}}
\if{latex}{\out{\hypertarget{method-acousticModel-simFit}{}}}
\subsection{Method \code{simFit()}}{
Simulate and Fit Acoustic Telemetry Data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{acousticModel$simFit(
  N,
  alpha,
  beta,
  q,
  gamma = NULL,
  emissionrate = 720,
  studyperiod,
  mu = NULL,
  absorbingstates = NULL,
  startbbox = NULL,
  formula = NULL,
  control = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{N}}{Number of animals.}

\item{\code{alpha}}{Diffusion rate.}

\item{\code{beta}}{Advection rates.}

\item{\code{q}}{Detection probability.}

\item{\code{gamma}}{Centre of attraction.}

\item{\code{emissionrate}}{Rate that the tags send a signal for detection.}

\item{\code{studyperiod}}{Time period of the study.}

\item{\code{mu}}{Rate of mortality per unit time.}

\item{\code{absorbingstates}}{Vector of states indicating which locations the animal will not leave if enters.}

\item{\code{startbbox}}{Vector indicating an area that the animal taggging is taking place.}

\item{\code{formula}}{Formula to relate covariates to advection (optional).}

\item{\code{control}}{List of object controls which include tolerance, rescale_freq, Nmax, uniformization, trace that controls the RTMB function \code{expAv}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Run simulation and then fit using \code{nlminb}.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-acousticModel-fitMove"></a>}}
\if{latex}{\out{\hypertarget{method-acousticModel-fitMove}{}}}
\subsection{Method \code{fitMove()}}{
Simulate and Fit Movement Using Telemetry Data on a fixed Interval.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{acousticModel$fitMove(
  N,
  delta_t,
  alpha,
  beta,
  gamma = NULL,
  emissionrate = 720,
  studyperiod,
  mu = NULL,
  absorbingstates = NULL,
  startbbox = NULL,
  formula = NULL,
  control = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{N}}{Number of animals.}

\item{\code{delta_t}}{Time between detections of the telemetry data.}

\item{\code{alpha}}{Diffusion rate.}

\item{\code{beta}}{Advection rates.}

\item{\code{gamma}}{Centre of attraction.}

\item{\code{emissionrate}}{Rate that the tags send a signal for detection.}

\item{\code{studyperiod}}{Time period of the study.}

\item{\code{mu}}{Rate of mortality per unit time.}

\item{\code{absorbingstates}}{Vector of states indicating which locations the animal will not leave if enters.}

\item{\code{startbbox}}{Vector indicating an area that the animal taggging is taking place.}

\item{\code{formula}}{Formula to relate covariates to advection (optional).}

\item{\code{control}}{List of object controls which include tolerance, rescale_freq, Nmax, uniformization, trace that controls the RTMB function \code{expAv}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Run simulation only if simulation hasn't been run. Then process data as fixed interval telemetry and then fit using \code{nlminb}.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-acousticModel-buildModel"></a>}}
\if{latex}{\out{\hypertarget{method-acousticModel-buildModel}{}}}
\subsection{Method \code{buildModel()}}{
Build Model with Joint Likelihood Gradient
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{acousticModel$buildModel(
  alpha,
  beta,
  q,
  gamma = NULL,
  emissionrate = 720,
  studyperiod,
  mu = NULL,
  control = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{Diffusion rate.}

\item{\code{beta}}{Advection rates.}

\item{\code{q}}{Detection probability.}

\item{\code{gamma}}{Centre of attraction.}

\item{\code{emissionrate}}{Rate that the tags send a signal for detection.}

\item{\code{studyperiod}}{Time period of the study.}

\item{\code{mu}}{Rate of mortality per unit time.}

\item{\code{control}}{List of object controls which include tolerance, rescale_freq, Nmax, uniformization, trace that controls the RTMB function \code{expAv}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Build negative log likelihood and gradient to be used for optimization.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-acousticModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-acousticModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{acousticModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
