% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{calc_states}
\alias{calc_states}
\title{Calculate state ID}
\usage{
calc_states(self, locs)
}
\arguments{
\item{self}{R6 object containing statespace and receivers.}

\item{Locs}{Dataframe of locations x,y.}
}
\description{
Calculate state ID
}
\details{
Compute which state a location is within.
}
