% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{calcLimit}
\alias{calcLimit}
\title{Calculate Limiting Probabilities}
\usage{
calcLimit(self, alpha, beta, gamma, mu)
}
\arguments{
\item{self}{R6 object containing statespace and receivers.}

\item{alpha}{Diffusion rate.}

\item{beta}{Advection rates.}

\item{gamma}{Centre of attraction.}

\item{mu}{Mortality rate.}
}
\value{
Vector of limiting probabilities, assuming that a steady state solution exists.
}
\description{
Calculate Limiting Probabilities
}
\details{
Solve for limiting distribution pi, as pi \%*\% Q = 0, and sum(pi) = 1.
}
