% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{compute_expected_time}
\alias{compute_expected_time}
\title{Compute Expected time between two points}
\usage{
compute_expected_time(
  self,
  deltat = 0.1,
  tstart = 0,
  tend = 1,
  s_init = NULL,
  s_end = NULL,
  alpha,
  beta,
  gamma,
  mu,
  q,
  emissionrate
)
}
\arguments{
\item{self}{R6 object containing statespace and receivers.}

\item{deltat}{Time steps to discretize}

\item{tstart}{Time to start computation.}

\item{tend}{Time to end computation.}

\item{s_init}{Initial location of the animal as a vector e.g. (0,0,0,1).}

\item{s_end}{Final location of the animal as a vector e.g. (0,0,0,1).}

\item{alpha}{Diffusion rate.}

\item{beta}{Advection rates.}

\item{gamma}{Centre of attraction.}

\item{mu}{Mortality rate.}
}
\value{
Vector of limiting probabilities, assuming that a steady state solution exists.
}
\description{
Compute Expected time between two points
}
\details{
Solve for limiting distribution pi, as pi \%*\% Q = 0, and sum(pi) = 1.
}
