% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{compute_observed_path}
\alias{compute_observed_path}
\title{Compute observed path}
\usage{
compute_observed_path(
  self,
  id = 1,
  deltat = 0.1,
  tend = NULL,
  alpha,
  beta,
  gamma,
  mu,
  q,
  emissionrate
)
}
\arguments{
\item{self}{R6 object containing statespace and receivers.}

\item{id}{Identity of animal to generate information.}

\item{deltat}{Time steps to discretize}

\item{tend}{Time to end computation.}

\item{alpha}{Diffusion rate.}

\item{beta}{Advection rates.}

\item{gamma}{Centre of attraction.}

\item{mu}{Mortality rate.}

\item{q}{Detection probability.}

\item{emissionrate}{Rate that the tags produce cues.}
}
\value{
Information needed for plotting observed paths.
}
\description{
Compute observed path
}
