% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expAdv_cpp}
\alias{expAdv_cpp}
\title{Compute expAv'}
\usage{
expAdv_cpp(A, dv, tol, renorm_freq, trans = FALSE)
}
\arguments{
\item{A}{A sparse matrix from R of \code{class(A) == "dgCMatrix"}.}

\item{dv}{Matrix of dv/dtheta for expAv', with nrows = nrow(A) and ncols = length(theta) to compute exp(A)*dv.}

\item{tol}{Tolerance default = 1e-8.}

\item{renorm_freq}{Renoramlization frequency to avoid computational overload from rho^n/n!.}

\item{trans}{Logical to confirm if we should do transpose of A.}
}
\description{
Compute expAv'
}
\details{
Implements Sherlock 2021 uniformization method, simply a version of expAv that allows for v to be a matrix.
}
