% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expAv_approx_cpp}
\alias{expAv_approx_cpp}
\title{Compute Linear approximation of expAv and gradient of expAv jointly}
\usage{
expAv_approx_cpp(A, v, lambda, tol)
}
\arguments{
\item{A}{A sparse matrix from R of \code{class(A) == "dgCMatrix"}.}

\item{v}{Vector (column vector) of to do exp(A)*v.}

\item{lambda}{Vector (column vector) of detection rate.}

\item{tol}{Tolerance default = 1e-8.}

\item{trans}{Logical to confirm if we should do transpose of A.}
}
\description{
Compute Linear approximation of expAv and gradient of expAv jointly
}
\details{
Computes joint expAv and the gradient as implemented in 'Differentiated uniformization: a new method for inferring Markov chains on combinatorial state spaces including stochastic epidemic models'.
}
