% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expAv_cpp}
\alias{expAv_cpp}
\title{Compute expAv}
\usage{
expAv_cpp(A, v, tol, renorm_freq, trans = FALSE)
}
\arguments{
\item{A}{A sparse matrix from R of \code{class(A) == "dgCMatrix"}.}

\item{v}{Vector (column vector) of to do exp(A)*v}

\item{tol}{Tolerance default = 1e-8.}

\item{renorm_freq}{Renoramlization frequency to avoid computational overload from rho^n/n!.}

\item{trans}{Logical to confirm if we should do transpose of A.}
}
\description{
Compute expAv
}
\details{
Implements Sherlock 2021 uniformization method.
}
