% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expAv_gr_cpp}
\alias{expAv_gr_cpp}
\title{Compute expAv and gradient of expAv jointly}
\usage{
expAv_gr_cpp(A, dA, v, tol, renorm_freq, row_indx, col_ptr)
}
\arguments{
\item{A}{A sparse matrix from R of \code{class(A) == "dgCMatrix"}.}

\item{dA}{A matrix that has row of non-zero gradient of A and columns for each parameter.}

\item{v}{Vector (column vector) of to do exp(A)*v}

\item{tol}{Tolerance default = 1e-8.}

\item{renorm_freq}{Renoramlization frequency to avoid computational overload from rho^n/n!.}

\item{row_indx}{Row index from column-sparse matrix "dgCMatrix" \code{A@i}.}

\item{col_ptr}{Column pointer from column-sparse matrix "dgCMatrix" \code{A@p}.}

\item{trans}{Logical to confirm if we should do transpose of A.}
}
\description{
Compute expAv and gradient of expAv jointly
}
\details{
Computes joint expAv and the gradient as implemented in 'Differentiated uniformization: a new method for inferring Markov chains on combinatorial state spaces including stochastic epidemic models'.
}
