% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{extend_detectors}
\alias{extend_detectors}
\title{Extend detector detection states}
\usage{
extend_detectors(self, dist = NULL)
}
\arguments{
\item{self}{R6 Acoustic telemetry object.}

\item{dist}{Distance that a detector can observe for changing resolution.}
}
\description{
Extend the maximum distance that even detection exists on a detector for a higher resolution grid.
}
