% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{extractControls}
\alias{extractControls}
\title{Extract Default Values}
\usage{
extractControls(controlValue, defaultValue)
}
\arguments{
\item{controlValue}{Any value which includes NULL.}

\item{defaultValue}{A set, non NULL value to return if controlValue is NULL.}
}
\value{
controlValue if it is not NULL, or defaultValue.
}
\description{
Extract Default Values
}
