% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_multiply}
\alias{fast_multiply}
\title{Fast Multiply of multiple sparse matrix vector.}
\usage{
fast_multiply(dA, term, row_indx, col_ptr)
}
\arguments{
\item{dA}{Dense matrix with each column the non-zero values of a sparse matrix.}

\item{term}{Vector to multiply against matrix.}

\item{row_indx}{Row index from column-sparse matrix "dgCMatrix" \code{A@i}.}

\item{col_ptr}{Column pointer from column-sparse matrix "dgCMatrix" \code{A@p}.}
}
\description{
Fast Multiply of multiple sparse matrix vector.
}
