% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{generator_design_gr}
\alias{generator_design_gr}
\title{Compute the advection-diffusion covariate space}
\usage{
generator_design_gr(self, formula)
}
\arguments{
\item{self}{R6 object containing statespace and receivers.}

\item{formula}{Formula (must exclude intercept) for which coviarate to include.}
}
\description{
Compute the advection-diffusion covariate space
}
\details{
Generate a design matrix that is computed as the central difference of the covariates for each direction Left/Right and Up/Down.
}
