% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{initCheck}
\alias{initCheck}
\title{Check initial values}
\usage{
initCheck(self, alpha, beta, gamma = NULL, verbose = FALSE)
}
\arguments{
\item{self}{R6 object for fitting Acoustic telemetry data.}

\item{alpha}{Difussion parameter in movement process.}

\item{beta}{Advection parameter relating to covariates in movement process. If OU process is included, the final term relates to the OU process.}

\item{gamma}{Vector of length 2, of a point of attraction for the animals in an OU process (default = NULL). Not included when NULL.}

\item{verbose}{Whether or not to print a warning if values are not valid.}
}
\description{
Check initial values
}
\details{
Test if alpha^2 - dx*(max(abs(beta \%*\% X))) is negative.
}
