% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{initValues}
\alias{initValues}
\title{Initialize random starting values}
\usage{
initValues(self, pars)
}
\arguments{
\item{self}{R6 object for fitting Acoustic telemetry data.}

\item{pars}{vector from a fitted object, or RTMB object parameters.}
}
\value{
pars vector ready for fitting.
}
\description{
Initialize random starting values
}
