% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inplace_transpose_sparse}
\alias{inplace_transpose_sparse}
\title{Overwrite a sparse matrix with its transpose}
\usage{
inplace_transpose_sparse(X)
}
\arguments{
\item{X}{A sparse matrix from R}
}
\description{
Overwrite a sparse matrix with its transpose
}
\details{
Basic function that forces a transpose on a \code{arma::sp_mat} matrix without generating two full in memory.
}
