% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.r
\name{make_ad_fun_mmpp}
\alias{make_ad_fun_mmpp}
\title{Build Likelihood for Markov modulated Poisson process}
\usage{
make_ad_fun_mmpp(
  self,
  alpha,
  beta,
  q,
  gamma = NULL,
  mu = NULL,
  control = list()
)
}
\arguments{
\item{self}{R6 object containing statespace and receivers.}

\item{alpha}{Diffusion parameter}

\item{beta}{Advection parameters, final value relates to OU process if gamma values provided.}

\item{q}{Detection rate within a state.}

\item{gamma}{Two values relating to the point of attraction in space (default = NULL).}

\item{mu}{Mortality rate per unit time (default = NULL).}

\item{control}{List of control values (default = list()).}
}
\description{
Build Likelihood for Markov modulated Poisson process
}
\details{
Compute the negative log likelihood for the Markov modulated Poisson process. Defines an AD model object as part of the \code{self} R6 object.
}
