% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.r
\name{make_expav_atomic}
\alias{make_expav_atomic}
\title{Make expAv Atomic for memory efficiency}
\usage{
make_expav_atomic(
  self,
  alpha,
  beta,
  q,
  mu = NULL,
  gamma = NULL,
  control = list()
)
}
\arguments{
\item{self}{R6 object containing statespace and receivers.}

\item{alpha}{Diffusion parameter}

\item{beta}{Advection parameters, final value relates to OU process if gamma values provided.}

\item{q}{Detection rate within a state.}

\item{mu}{Mortality rate per unit time (default = NULL).}

\item{gamma}{Two values relating to the point of attraction in space (default = NULL).}

\item{control}{List of values to input in the \code{expAv} function (default = list()).}
}
\value{
Atomic function of expAv.
}
\description{
Make expAv Atomic for memory efficiency
}
\details{
Builds the atomic function to compute expAv with memory efficiency. If \code{control$tranpose} is TRUE then we compute vexp(A) or FALSE we compute exp(A)v.
\code{control$mmpp} will include the detection process or will just compute movement process.
}
