% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.r
\name{make_generator}
\alias{make_generator}
\title{Make CTMC Generator}
\usage{
make_generator(self, alpha, beta, mu = NULL, gamma = NULL, nstates, dx)
}
\arguments{
\item{self}{R6 object containing statespace and receivers.}

\item{alpha}{Diffusion parameter}

\item{beta}{Advection parameters, final value relates to OU process if gamma values provided.}

\item{mu}{Mortality rate per unit time (default = NULL).}

\item{gamma}{Two values relating to the point of attraction in space (default = NULL).}

\item{nstates}{Number of states in CTMC.}

\item{dx}{Grid cell width in each state.}
}
\value{
Generator Matrix.
}
\description{
Make CTMC Generator
}
\details{
Build the generator matrix for the animal movement CTMC.
}
