% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.r
\name{make_negll_gr}
\alias{make_negll_gr}
\title{Fit Acoustic Model with joint loglikelihood and gradient evaluation}
\usage{
make_negll_gr(self, alpha, beta, q, gamma = NULL, mu = NULL, control = list())
}
\arguments{
\item{self}{R6 object containing statespace and receivers.}

\item{alpha}{Diffusion parameter}

\item{beta}{Advection parameters, final value relates to OU process if gamma values provided.}

\item{q}{Detection rate within a state.}

\item{gamma}{Two values relating to the point of attraction in space (default = NULL).}

\item{mu}{Mortality rate per unit time (default = NULL).}

\item{control}{List of control values (default = list()).}
}
\value{
nlminb fitted object.
}
\description{
Fit Acoustic Model with joint loglikelihood and gradient evaluation
}
\details{
Compute the negative log likelihood for the Markov modulated Poisson process and fit it using \code{nlminb}.
}
