% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.r
\name{make_Q_rtmb}
\alias{make_Q_rtmb}
\title{Make Observed Generator Matrix in AD for gradients}
\usage{
make_Q_rtmb(self, alpha, beta, q, mu = NULL, gamma = NULL, control = list())
}
\arguments{
\item{self}{R6 object containing statespace and receivers.}

\item{alpha}{Diffusion parameter}

\item{beta}{Advection parameters, final value relates to OU process if gamma values provided.}

\item{q}{Detection rate within a state.}

\item{mu}{Mortality rate per unit time (default = NULL).}

\item{gamma}{Two values relating to the point of attraction in space (default = NULL).}

\item{control}{List of values to input in the \code{expAv} function (default = list()).}
}
\value{
Taped functin of Q for a input of parameters.
}
\description{
Make Observed Generator Matrix in AD for gradients
}
\details{
Builds the atomic function to compute the generator matrix in MMPP.
}
