% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot_limit}
\alias{plot_limit}
\title{Plot surfaces}
\usage{
plot_limit(self, alpha, beta, gamma, mu, layer, label, quantiles = NULL)
}
\arguments{
\item{self}{R6 object containing statespace and receivers.}

\item{alpha}{Diffusion rate.}

\item{beta}{Advection rates.}

\item{gamma}{Centre of attraction.}

\item{mu}{Mortality rate.}

\item{quantiles}{Quantiles to plot for the limting distribution (default = c(0.95, 0.8, 0.5))}

\item{Layer}{Habitat layer name.}

\item{Label}{to name the habitat in the plot.}
}
\description{
Plot surfaces
}
\details{
Plot habitat variable and the limiting distribution as a contour.
}
