% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot_observed_path}
\alias{plot_observed_path}
\title{Plot Expected Time along with Path}
\usage{
plot_observed_path(
  self,
  id = 1,
  deltat = 0.1,
  tend = NULL,
  alpha,
  beta,
  gamma,
  mu,
  q,
  emissionrate
)
}
\arguments{
\item{self}{R6 object containing statespace and receivers.}

\item{deltat}{Time steps to discretize}

\item{tend}{Time to end computation.}

\item{alpha}{Diffusion rate.}

\item{beta}{Advection rates.}

\item{gamma}{Centre of attraction.}

\item{mu}{Mortality rate.}

\item{tstart}{Time to start computation.}

\item{s_init}{Initial location of the animal as a vector e.g. (0,0,0,1).}

\item{s_end}{Final location of the animal as a vector e.g. (0,0,0,1).}
}
\value{
Vector of limiting probabilities, assuming that a steady state solution exists.
}
\description{
Plot Expected Time along with Path
}
\details{
Solve for limiting distribution pi, as pi \%*\% Q = 0, and sum(pi) = 1.
}
