% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot_path}
\alias{plot_path}
\title{Plot Expected Path}
\usage{
plot_path(
  self,
  deltat = 0.1,
  tstart = 0,
  tend = 1,
  s_init = NULL,
  alpha,
  beta,
  gamma,
  mu,
  expected = TRUE
)
}
\arguments{
\item{self}{R6 object containing statespace and receivers.}

\item{alpha}{Diffusion rate.}

\item{beta}{Advection rates.}

\item{gamma}{Centre of attraction.}

\item{mu}{Mortality rate.}
}
\value{
Vector of limiting probabilities, assuming that a steady state solution exists.
}
\description{
Plot Expected Path
}
\details{
Solve for limiting distribution pi, as pi \%*\% Q = 0, and sum(pi) = 1.
}
