% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{process_data}
\alias{process_data}
\title{Process Acoustic Telemetry Data for Modelling}
\usage{
process_data(
  self,
  detections = NULL,
  releases = NULL,
  known_fates = NULL,
  absorbing_states = NULL,
  studyperiod = NULL
)
}
\arguments{
\item{self}{R6 object containing statespace and receivers.}

\item{detections}{Detection history dataframe that includes x,y coordinates of detectors (default = NULL). Default implies simulated data.}

\item{releases}{Dataframe of where the animals were released (Default NULL).}

\item{known_fates}{Dataframe of where the animals ended up (Default NULL).}

\item{absorbing_states}{Vector of which states are considered absorbing, such as entering a river system (default = NULL).}
}
\description{
Process Acoustic Telemetry Data for Modelling
}
