% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{process_data_movement}
\alias{process_data_movement}
\title{Process Telemetry Data for Modelling}
\usage{
process_data_movement(self, delta_t = 1, absorbing_states = NULL)
}
\arguments{
\item{self}{R6 object containing statespace and receivers.}

\item{delta_t}{Time increment.}

\item{absorbing_states}{Vector of which states are considered absorbing, such as entering a river system (default = NULL).}
}
\description{
Process Telemetry Data for Modelling
}
