% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{process_states_detectors}
\alias{process_states_detectors}
\title{Process State Space}
\usage{
process_states_detectors(self, statespace, detectors)
}
\arguments{
\item{self}{R6 object containing statespace and receivers.}

\item{statespace}{Dataframe of grid cells with x,y values that define the discrete space and includes any covariates.}

\item{detectors}{Dataframe of detector locations.}
}
\description{
Process State Space
}
\details{
Process the locations in both state space and receivers to discrete space. Annotate each with which state is left, right, up and down.
}
