% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{reList}
\alias{reList}
\title{Re-List parameters}
\usage{
reList(pars)
}
\arguments{
\item{pars}{vector from a fitted object, or RTMB object parameters.}
}
\value{
List of parameters back on the real scale.
}
\description{
Re-List parameters
}
