% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{simulate_gmrf}
\alias{simulate_gmrf}
\title{Simulate a Gaussian Markov Random Field}
\usage{
simulate_gmrf(x, y, nhabitat = 1, kappa = 0.95)
}
\arguments{
\item{x}{Vector of centroids of states along x axis.}

\item{y}{Vector of centroids of states along y axis.}

\item{nhabitat}{Number of habitat variables to simulate.}

\item{kappa}{Vector of kappa values to simulate the amount of habitat.}
}
\description{
Simulate a Gaussian Markov Random Field
}
