% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{simulate_mmpp}
\alias{simulate_mmpp}
\title{Simulate a Marokov Modulate Poisson process for acoustic telemetry.}
\usage{
simulate_mmpp(
  self,
  N,
  alpha,
  beta,
  q = 0.3,
  gamma = NULL,
  startbbox = NULL,
  mu = NULL
)
}
\arguments{
\item{self}{R6 object containing statespace and receivers.}

\item{N}{Number of animals to simulate.}

\item{alpha}{Difussion parameter in movement process.}

\item{beta}{Advection parameter relating to covariates in movement process. If OU process is included, the final term relates to the OU process.}

\item{q}{Detection rate per unit time within a state (default = 0.3).}

\item{gamma}{Vector of length 2, of a point of attraction for the animals in an OU process (default = NULL). Not included when NULL.}

\item{startbbox}{bounding box as a vector with values \code{c('xmin', 'ymin', 'xmax', 'ymax')}}

\item{mu}{Mortality rate per unit time (default = NULL). No mortality assumed when NULL.}
}
\description{
Simulate a Marokov Modulate Poisson process for acoustic telemetry.
}
