% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.r
\name{vertibi_algorithm}
\alias{vertibi_algorithm}
\title{Vertibi Algorithm}
\usage{
vertibi_algorithm(
  self,
  alpha,
  beta,
  q,
  gamma = NULL,
  mu = NULL,
  deltat = NULL,
  control = list()
)
}
\arguments{
\item{self}{R6 object containing statespace and receivers.}

\item{alpha}{Diffusion parameter}

\item{beta}{Advection parameters, final value relates to OU process if gamma values provided.}

\item{q}{Detection rate within a state.}

\item{gamma}{Two values relating to the point of attraction in space (default = NULL).}

\item{mu}{Mortality rate per unit time (default = NULL).}

\item{deltat}{Time step to use (default = NULL).}

\item{control}{List of control values (default = list()).}

\item{tstart}{Start time of algorithm (default = 0).}

\item{tend}{End time of algoirthm (default = NULL).}
}
\value{
List of \code{forwardprob} an array for each animal each time step the probability in each location,  \code{backwardprob} an array for each animal each step the probability in each location,
\code{times} vector of each time step, and \code{path} matrix of the expected path for each animal.
}
\description{
Vertibi Algorithm
}
\details{
Run the forward algorithm for each animal and detection, then the backward algorithm and then finally the Vertibi algorithm to compute the expected path.
}
